Break awesome out into its own migration
You'll notice I put a fake date in `db/awesome.plan`... I had to do this to make sure they migrated in the correct order. You can use this as a reference for how to set it up so one migration requires another.

In this example, `python` requires `awesome` (see `db/awesome.plan`)